save RAM

+++ MashController_pio/include/Sensor.h
@@ -15,7 +15,7 @@ extern bool sensorOk;
 extern float lastGoodTemp;
 extern bool conversionInProgress;   // exposed so /rediscoverSensor can check it
 
-extern float tempHistory[HISTORY_SIZE];
+extern int16_t tempHistory[HISTORY_SIZE];  // holds the last HISTORY_SIZE readings, in t/100 C units
 extern int histIndex;
 
 void sensorInit();     // call once from setup(), after Serial.begin()
